refactor: SnackBar 공통 컴포넌트로 전환 및 적용
