+++ src/pages/MyProfile/index.tsx
@@ -14,6 +14,7 @@ import {
   QUERY_KEYS,
 } from '@/pages/setting/my-profile';
 import { useProfileImageResetter, useProfileInfoUpdater } from './mutations';
+import SnackBar from '@/src/components/SnackBar';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 type MY_PROFILE_INPUT_TYPE = 'description' | 'interests';
@@ -239,22 +240,6 @@ function Tag({ children, onClick }: TagProps) {
   );
 }
 
-// SnackBar Component
-interface SnackBarProps {
-  text: string;
-}
-
-function SnackBar({ text }: SnackBarProps) {
-  const [isVisible, setIsVisible] = useState(false);
-  useEffect(() => {
-    setIsVisible(true);
-    setTimeout(() => {
-      setIsVisible(false);
-    }, 3000);
-  }, []);
-  return <SnackBarWrapper isVisible={isVisible}>{text}</SnackBarWrapper>;
-}
-
 export default MyProfile;
 
 const Wrapper = styled.div`
@@ -340,23 +325,6 @@ const WardTitle = styled.div`
   margin: 30px 0 16px 0;
   color: ${({ theme }) => theme.color.gray.Gray400};
   ${typography.Body_Medium_14}
-`;
-
-const SnackBarWrapper = styled.div<{ isVisible: boolean }>`
-  position: absolute;
-  bottom: 104px;
-  left: calc(50% - 115px);
-  width: 230px;
-  height: 38px;
-  padding: 10px 32px;
-  color: ${({ theme }) => theme.color.basic.White};
-  text-align: center;
-  /* FIXME: 디자인시스템에 없는 색상 디자인팀에 문의 필요 */
-  background-color: rgba(37, 37, 37, 0.8);
-  border-radius: 4px;
-  opacity: ${({ isVisible }) => (isVisible ? 0.95 : 0)};
-  transition: opacity 1s;
-  ${typography.Caption1_Regular_13}
 `;
 
 const FileInput = styled.input`
